+++ sudokuextract/imgproc/geometry.py
@@ -17,14 +17,15 @@ from __future__ import absolute_import
 from operator import attrgetter
 
 import numpy as np
-from skimage.morphology import binary_closing, binary_erosion, binary_opening, disk
+from skimage.morphology import binary_opening
+from skimage.filters import gaussian_filter
 from skimage.transform import warp, ProjectiveTransform
 from skimage.measure import find_contours
 
 from sudokuextract.imgproc.binary import to_binary_otsu, to_binary_adaptive
 
 
-def warp_image(corner_points, image):
+def warp_image_by_corner_points_projection(corner_points, image):
     """Given corner points of a Sudoku, warps original selection to a square image.
 
     :param corner_points:
@@ -49,7 +50,7 @@ def warp_image(corner_points, image):
 
     tr = ProjectiveTransform()
     tr.estimate(dst, src)
-    
+
     return warp(image, tr, output_shape=(L, L))
 
 
@@ -82,52 +83,18 @@ def warp_image_by_interp_borders(edges, image):
     d_right_edge = np.linalg.norm(right_edge[0, :] - right_edge[-1, :])
 
     d = int(np.ceil(max([d_top_edge, d_bottom_edge, d_left_edge, d_right_edge])))
-    wi = warp(image, _mapping_fcn, output_shape=(d, d))
-
-    return wi
+    return warp(image, _mapping_fcn, output_shape=(d, d))
 
 
-def warp_image_by_projmap_borders(edges, image):
-    left_edge, top_edge, right_edge, bottom_edge = edges
-
-    left_edge = left_edge[::-1, :]
-    bottom_edge = bottom_edge[::-1, :]
-
-    d_top_edge = np.linalg.norm(top_edge[0, :] - top_edge[-1, :])
-    d_bottom_edge = np.linalg.norm(bottom_edge[0, :] - bottom_edge[-1, :])
-    d_left_edge = np.linalg.norm(left_edge[0, :] - left_edge[-1, :])
-    d_right_edge = np.linalg.norm(right_edge[0, :] - right_edge[-1, :])
-
-    d = int(np.ceil(max([d_top_edge, d_bottom_edge, d_left_edge, d_right_edge])))
-
-    left_edge_proj = np.zeros((len(left_edge), 2))
-    left_edge_proj[:, 1] = np.linspace(d - 1, 0.0, num=len(left_edge))
-
-    top_edge_proj = np.zeros((len(top_edge), 2))
-    top_edge_proj[:, 0] = np.linspace(0.0, d - 1, num=len(top_edge))
-
-    right_edge_proj = np.ones((len(right_edge), 2)) * (d - 1)
-    right_edge_proj[:, 1] = np.linspace(0.0, d - 1, num=len(right_edge))
-
-    bottom_edge_proj = np.ones((len(bottom_edge), 2)) * (d - 1)
-    bottom_edge_proj[:, 0] = np.linspace(d - 1, 0.0, num=len(bottom_edge))
-
-    dst = np.concatenate([left_edge, top_edge, right_edge, bottom_edge])
-    src = np.concatenate([left_edge_proj, top_edge_proj, right_edge_proj, bottom_edge_proj])
-
-    tr = ProjectiveTransform()
-    tr.estimate(src, dst)
-    wi = warp(image, tr, output_shape=(d, d))
-
-    return wi
-
-
-def split_image_into_sudoku_pieces_adaptive_global(image):
+def split_image_into_sudoku_pieces_adaptive_global(image, otsu_local=False, apply_gaussian=False):
     L = image.shape[0]
     d = int(np.ceil(L / 9))
     dd = d // 6
     output = []
-    image = to_binary_adaptive(image)
+    if apply_gaussian:
+        image = gaussian_filter(image, sigma=10.0)
+    if not otsu_local:
+        image = to_binary_adaptive(image)
     for k in range(9):
         this_row = []
         start_row_i = max([k * d - dd, 0])
@@ -136,32 +103,13 @@ def split_image_into_sudoku_pieces_adaptive_global(image):
             start_col_i = max([kk * d - dd, 0])
             stop_col_i = min([(kk + 1) * d + dd, L])
             i = image[start_row_i:stop_row_i, start_col_i:stop_col_i].copy()
+            if otsu_local:
+                i = to_binary_otsu(i)
             i = binary_opening(i)
             i = to_binary_otsu(i)
             this_row.append(i)
         output.append(this_row)
-    return output
-
-
-def split_image_into_sudoku_pieces_otsu_local(image):
-    L = image.shape[0]
-    d = int(np.ceil(L / 9))
-    dd = d // 6
-    output = []
-    for k in range(9):
-        this_row = []
-        start_row_i = max([k * d - dd, 0])
-        stop_row_i = min([(k + 1) * d + dd, L])
-        for kk in range(9):
-            start_col_i = max([kk * d - dd, 0])
-            stop_col_i = min([(kk + 1) * d + dd, L])
-            i = image[start_row_i:stop_row_i, start_col_i:stop_col_i].copy()
-            i = to_binary_otsu(i)
-            i = binary_opening(i)
-            i = to_binary_otsu(i)
-            this_row.append(i)
-        output.append(this_row)
-    return output
+    return output, image
 
 
 def get_contours(img, contour_param=0.8):